% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_change_cumu.R
\name{calc_change_cumu}
\alias{calc_change_cumu}
\title{Calculate station- or site-level cumulative change for SET or MH data}
\usage{
calc_change_cumu(data, level = "station")
}
\arguments{
\item{data}{A data frame. A data frame of raw SET or MH data. See details
below for requirements.}

\item{level}{A string (optional). Level at which to calculate rates of
surface elevation change or vertical accretion. One of:
* `"station"`: (default) station-level rates of surface elevation change.
* `"site"`: site-level rates of surface elevation change.}
}
\value{
For SET data, returns a data frame of station- or site-level
 cumulative surface elevation change. For MH data, returns a data frame of
 station- or site-level vertical accretion.
}
\description{
This function takes a data frame of raw SET or MH data and calculates
cumulative surface elevation change (for SET data) or vertical accretion (for
MH data) on each measurement date at either the station- or site-level.
}
