% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_set_rates.R
\name{calc_set_rates}
\alias{calc_set_rates}
\title{Calculate SET-level (i.e., station-level) linear rates of elevation change}
\usage{
calc_set_rates(data)
}
\arguments{
\item{data}{a data frame with one row per pin reading, and the following columns, named exactly: event_date_UTC, network_code, park_code, site_name, station_code, SET_direction, pin_position, pin_height_mm}
}
\value{
A data frame with linear regressions models, rates of change, standard errors, and confidence intervals for each SET.
}
\description{
Calculate SET-level (i.e., station-level) linear rates of elevation change
}
\examples{
calc_set_rates(example_sets)

}
