% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sea_level_data.R
\name{get_sea_level_data}
\alias{get_sea_level_data}
\title{Get park-specific SLR rates from NOAA tides and currents}
\usage{
get_sea_level_data(
  park,
  nauset = FALSE,
  trend_start_year = NULL,
  trend_end_year = NULL
)
}
\arguments{
\item{park}{A string (required). A capitalized four-letter park code. One of:
* `"COLO"`: Colonial National Historical Park: NOAA gauge 8638610.
* `"ASIS"`: Assateague Island National Seashore: NOAA gauge 8570283.
* `"GWMP"`: George Washington Memorial Parkway: NOAA gauge 8594900.
* `"NACE"`: National Capital Area East: NOAA gauge 8594900.
* `"GATE"`: Gateway National Recreation Area: NOAA gauge 8531680.
* `"FIIS"`: Fire Island National Seashore: NOAA gauge 8531680.
* `"CACO"`: Cape Cod National Seashore: NOAA gauge 8443970 or 8449130 (Nauset).
* `"BOHA"`: Boston Harbor Islands National Recreation Area: NOAA gauge 8443970.
* `"ACAD"`: Acadia National Park: NOAA gauge 8413320.}

\item{nauset}{Logical. Used to select the Nauset gauge for CACO; defaults to
FALSE.}

\item{trend_start_year}{An integer year (optional). The starting year that
you want to use to calculate SLR. If trend_start_year is not supplied
(default), the SLR calculation uses the earliest available date.}

\item{trend_end_year}{An integer year (optional). The ending year that you
want to use to calculate SLR. If trend_end_year is not supplied (default),
the SLR calculation uses the latest available date.}
}
\value{
A list with 2 data frames: The first data frame, `slr_data`, is the
  relative sea-level data downloaded from NOAA tides and current. The column
  "Monthly_MSL_m" gives the monthly mean sea-level in meters, whereas
  "Monthly_MSL_mm" is in millimeters. The second data frame, `slr_rate`,
  gives the calculated rate of sea-level rise in mm/yr (column "slr_rate"),
  standard error of the SLR rate (column "slr_rate_se"), lower confidence
  interval (column "lower_ci"), upper confidence interval (column
  "upper_ci"), minimum year of data used for calculating SLR (column
  "min_year"), and maximum year of data used for calculating SLR (column
  "max_year").
}
\description{
This function downloads NOAA tides and currents from the nearest gauge and
calculates a rate of SLR.
}
\details{
The rate of sea-level rise for each park is calculated using
  relative sea-level data (RSLR) from the nearest NOAA tide gauge. RSLR Data
  is downloaded from the NOAA tides and currents website
  https://tidesandcurrents.noaa.gov/.
}
\note{
Note that monthly mean sea level values are relative to the mean
  sea-level datum for each station.
}
\examples{
get_sea_level_data(park = "ASIS")

get_sea_level_data(park = "CACO", nauset = TRUE)

get_sea_level_data(park = "COLO", trend_start_year = 2008, trend_end_year = 2018)

}
