% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mh_data.R
\name{load_mh_data}
\alias{load_mh_data}
\title{Load MH data from database or saved file}
\usage{
load_mh_data(
  park = NULL,
  network_code = NULL,
  file_path = NULL,
  db_version = "production"
)
}
\arguments{
\item{park}{(optional); A 4 character park code that can be used to filter
results from the NPS SET database to a specific park. If not specified,
will return data for all parks.}

\item{network_code}{(optional); A 4 character network code that can be used to
filter results from the NPS SET database to a specific I&M network. If not
specified, will return data for all networks.}

\item{file_path}{(optional); If, instead of connecting to the NPS SET database,
you want to use a saved .csv or .xlsx file, the full file path to the saved
file. Must be table with one row per marker horizon reading, and the
following columns, named exactly: event_date_UTC, network_code, park_code,
site_name, marker_horizon_name, core_measurement_number,
core_measurement_depth_mm, established_date.}

\item{db_version}{(optional); If getting data from database, do you want to
connect to the production or development version of the database? Default
is "production".}
}
\value{
A data frame of raw MH data.
}
\description{
Load MH data from database or saved file
}
\details{
Note that you must be on VPN and have read access to the NPS I&M SET
  database for the database option to work.#'
}
\examples{

load_mh_data(park = "ASIS")

load_mh_data(network_code = "NCBN",  db_version = "development")

load_mh_data(file_path = "C:/Documents/Data/my_MH_data.csv")

}
