% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumu_arm.R
\name{plot_cumu_arm}
\alias{plot_cumu_arm}
\title{Make a graph of change over time by arm position}
\usage{
plot_cumu_arm(data, set = NULL, columns = 4, pointsize = 2, scales = "fixed")
}
\arguments{
\item{data}{Data frame with one row per pin reading, and the following columns, named exactly: event_date_UTC, network_code, park_code, site_name, station_code, SET_direction, pin_position, and pin_height_mm.}

\item{columns}{Number of columns you want in the faceted output; defaults to 4.}

\item{pointsize}{Size of points you want (goes into the `size` argument of `ggplot2::geom_point`); defaults to 2.}

\item{scales}{Do you want axis scales to be the same in all facets ("fixed") or to vary between facets "free_x" or "free_y" or "free" - goes into `scales` arg of `facet_wrap`; defaults to "free_y".}
}
\value{
a ggplot object
}
\description{
x-axis is date; y-axis is the average of the 9 pin heights' difference from baseline (first measurement) for each arm. One facet per SET/station id.
}
\examples{
plot_cumu_arm(example_sets)
plot_cumu_arm(example_sets, columns = 1, pointsize = 2)
}
