% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumu_site.R
\name{plot_cumu_site}
\alias{plot_cumu_site}
\title{Make a graph of change over time by site}
\usage{
plot_cumu_site(
  data,
  columns = 2,
  pointsize = 3.5,
  scales = "fixed",
  smooth = TRUE,
  lty_smooth = 5
)
}
\arguments{
\item{scales}{Do you want axis scales to be the same in all facets ("fixed") or to vary between facets "free_x" or "free_y" or "free" - goes into `scales` arg of `facet_wrap`; defaults to "free_y".}

\item{smooth}{Do you want a linear regression plotted on top? defaults to TRUE.}

\item{lty_smooth}{What type of line do you want the linear regression to plot as? 1 = solid; 2 and 5 = dashed; or any of the other line types available in `ggplot2`. defaults to dashed.}
}
\value{
a ggplot object
}
\description{
Make a graph of change over time by site
}
\examples{
plot_cumu_site(example_sets)
plot_cumu_site(example_sets, columns = 1, pointsize = 2, smooth = FALSE)
}
