% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_arm.R
\name{plot_hist_arm}
\alias{plot_hist_arm}
\title{Generate a histogram of pin readings by arm position}
\usage{
plot_hist_arm(data, columns = 4, scales = "free_y")
}
\arguments{
\item{data}{Data frame with one row per pin reading, and the following columns, named exactly: event_date_UTC, network_code, park_code, site_name, station_code, SET_direction, pin_position, and pin_height_mm.}

\item{columns}{Number of columns you want in the faceted output; defaults to 4.}

\item{scales}{Do you want axis scales to be the same in all facets ("fixed") or to vary between facets "free_x" or "free_y" or "free" - goes into `scales` arg of `facet_wrap`; defaults to "free_y".}
}
\value{
a ggplot object
}
\description{
Generate a histogram of pin readings by arm position
}
\examples{
plot_hist_arm(example_sets)
}
