% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incr_arm.R
\name{plot_incr_arm}
\alias{plot_incr_arm}
\title{Plot change between readings, by SET_direction (i.e., arm)}
\usage{
plot_incr_arm(
  data,
  set = NULL,
  threshold = 25,
  columns = 4,
  pointsize = 2,
  scales = "fixed"
)
}
\arguments{
\item{data}{Data frame with one row per pin reading, and the following columns, named exactly: event_date_UTC, network_code, park_code, site_name, station_code, SET_direction, pin_position, and pin_height_mm.}

\item{set}{optional SET (aka station) ID if you only want to look at one SET; default is to graph all SETs.}

\item{threshold}{Numeric value for red horizontal lines (at +/- this value); can be used for QAQC of arm_level incremental change; defaults to 25.}

\item{columns}{Number of columns you want in the faceted output; defaults to 4.}

\item{pointsize}{Size of points you want (goes into the `size` argument of `ggplot2::geom_point`); defaults to 2.}

\item{scales}{Do you want axis scales to be the same in all facets ("fixed") or to vary between facets "free_x" or "free_y" or "free" - goes into `scales` arg of `facet_wrap`; defaults to "fixed".}
}
\value{
a ggplot object
}
\description{
Plot change between readings, by SET_direction (i.e., arm)
}
\examples{
plot_incr_arm(example_sets)
plot_incr_arm(example_sets, threshold = 5, columns = 1)
plot_incr_arm(example_sets, set = "M11-1", threshold = 5)
}
