% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_site_rate_comps.R
\name{plot_site_rate_comps}
\alias{plot_site_rate_comps}
\title{Graphical comparison of Site-level elevation change rates}
\usage{
plot_site_rate_comps(
  data,
  plot_type = 2,
  error_bar_type = "se",
  site_ids = site_name,
  rates = site_rate,
  site_se = se_rate,
  site_ci_low = ci_low,
  site_ci_high = ci_high,
  comp1 = NULL,
  comp1_se = NULL,
  comp1_ci_low = NULL,
  comp1_ci_high = NULL,
  comp1_name = "Comp1 rate",
  comp2 = NULL,
  comp2_se = NULL,
  comp2_ci_low = NULL,
  comp2_ci_high = NULL,
  comp2_name = "Comp2 rate"
)
}
\arguments{
\item{data}{Either 1) a data frame with one row per pin reading, and the following columns, named exactly: event_date_UTC, network_code, park_code, site_name, station_code, SET_direction, pin_position, pin_height_mm; or 2) a user-supplied data frame of site-level elevation change rates with (at least) columns for site IDs, site-level rates, and either a) a single column with site-level rate std error, or b) 2 columns with the high and low site-level rate confidence intervals.}

\item{plot_type}{One of various combinations of points, error bars, or confidence intervals; default is 2. 1 = basic; points only; no error bars/confidence intervals; 2 = SEs/CIs for SET rates; 3 = SEs/CIs for SET rates and a rate for comparison (e.g. sea-level rise); 4 = SEs/CIs for SET rates and 2 different rates for comparisons.}

\item{error_bar_type}{For plot_types 2-4, choose display with either standard errors or 95% confidence intervals. 'se' = standard errors (default); 'confint' = 95% confidence intervals.}

\item{site_ids}{Column name containing unique site IDs or names. Defaults to site_name. The column name must be specified if using a user-supplied data frame of site-level rates.}

\item{rates}{Column name of numbers representing the linear estimates of site-level rates of elevation change. Defaults to site_rate. The column name must be specified if using a user-supplied data frame of site-level rates.}

\item{site_se}{Column name of numbers representing the standard errors for each site-level rate of elevation change. Defaults to se_rate. The column name must be specified if using a user-supplied data frame of site-level rates.}

\item{site_ci_low}{Column name of numbers representing the lower limit of the 95\% confidence interval for each site-level rate of elevation change. Defaults to ci_low. The column name must be specified if using a user-supplied data frame of site-level rates.}

\item{site_ci_high}{Column of numbers representing the upper limit of the 95\% confidence interval for each site-level rate of elevation change. Defaults to ci_high. The column name must be specified if using a user-supplied data frame of site-level rates.}

\item{comp1}{optional; Single number for comparing to site-level rates of change (e.g. long-term Sea Level Rise) - will appear as a line on the graph.}

\item{comp1_se}{optional; Single number representing 1 standard error for the point estimate 'comp1'. Required if comp1 is specified and error_bar_type = "se".}

\item{comp1_ci_low}{optional; Single number representing the lower limit of the 95\% confidence interval for the point estimate `comp1`. Required if comp1 is specified and error_bar_type = "confint".}

\item{comp1_ci_high}{optional; Single number representing the upper limit of the 95\% confidence interval for the point estimate `comp1`. Required if comp1 is specified and error_bar_type = "confint".}

\item{comp1_name}{optional; Label for comp1 value (e.g. "Local, long-term SLR").}

\item{comp2}{optional; Single number for comparing to site-level rates of change (e.g. 19-year water level change) - will appear as a line on the graph.}

\item{comp2_se}{optional; Single number representing 1 standard error for the point estimate 'comp2'. Required if comp2 is specified and error_bar_type = "se".}

\item{comp2_ci_low}{optional; Single number representing the lower limit of the 95\% confidence interval for the point estimate `comp2`. Required if comp2 is specified and error_bar_type = "confint".}

\item{comp2_ci_high}{optional; a single number representing the upper limit of the 95\% confidence interval for the point estimate `comp2`. Required if comp2 is specified and error_bar_type = "confint".}

\item{comp2_name}{optional; Label for comp2 value (e.g. "19-yr water level change").}
}
\value{
a ggplot object
}
\description{
Graphical comparison of Site-level elevation change rates
}
\examples{

plot_site_rate_comps(data = example_sets,
                plot_type = 1)

plot_site_rate_comps(data = example_sets,
               plot_type = 2,
               error_bar_type = "confint")

plot_site_rate_comps(
    data = example_sets,
    plot_type = 3,
    error_bar_type = "se",
    comp1 = 3.5,
    comp1_se = 0.2,
    comp1_name = "Local, long-term SLR")


# Example with a user-supplied data frame of site-level rates
example_rates <- data.frame("site_id" = c("site1", "site2", "site3"),
                            "set_rate" = c(3.2, 4.0, 5.4),
                            "ci_low" = c(3.0, 3.2, 5.2),
                            "ci_high" = c(3.4, 4.8, 5.6))


plot_site_rate_comps(data = example_rates,
                plot_type = 4,
                error_bar_type = "confint",
                site_ids = set_id,
                rates = site_rate,
                set_ci_low = ci_low,
                set_ci_high = ci_high,
                comp1 = 3.5,
                comp1_ci_low = 3.3,
                comp1_ci_high = 3.7,
                comp1_name = "Local, long-term SLR",
                comp2 = 4.0,
                comp2_ci_low = 3.6,
                comp2_ci_high = 4.4,
                comp2_name = "19-yr water level change")

}
