% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_set_mh_data.R
\name{write_set_mh_data}
\alias{write_set_mh_data}
\title{Save data to a csv file}
\usage{
write_set_mh_data(
  data,
  dest_folder = NULL,
  create_folders = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{data}{data frame. Can be a data frame of raw SET/MH data produced by
`load_set_data` or `load_mh_data`, a data frame of cumulative change
produced by `calc_change_cumu`, or a data frame of calculated linear rates
produced by `calc_linear_rates`.}

\item{dest_folder}{string (optional). The folder where you want the file to
be saved. Defaults to the current working directory.}

\item{create_folders}{boolean (TRUE/FALSE). If the folder specified in
`dest_folder` doesn't exist, do you want to create it? Defaults to FALSE.}

\item{overwrite}{boolean (TRUE/FALSE). If a file with the same name already
exists in `dest_folder`, do you want to overwrite it? Defaults to FALSE.}
}
\value{
Saves a data frame to a csv. The file name will be the name of the
  data frame supplied to `data` suffixed with the current date e.g.
  "colo_set_data_2025-06-04.csv".
}
\description{
This function saves a data frame to a csv file. Can be a data frame of
raw SET/MH data produced by `load_set_data` or `load_mh_data`, a data frame
of cumulative change produced by `calc_change_cumu`, or a data frame of
calculated linear rates produced by `calc_linear_rates`.
}
\examples{
\dontrun{
# Load SET data from COLO and write to csv

colo_set_data <- load_set_data(park = "COLO")

write_set_mh_data(
    data = colo_set_data,
    dest_folder = "C:/Documents/SET_data",
    create_folders = TRUE,
    overwrite = FALSE
    )

# Calculate station-level cumulative change and write to csv in the current working directory

colo_cumu_change <- calc_change_cumu(colo_set_data, level = "station")

write_set_mh_data(
    data = colo_cumu_change,
    overwrite = TRUE)

# Calculate site-level linear rates of change and write to csv in the current working directory

colo_site_set_rates <- calc_linear_rates(colo_set_data, level = "site")

write_set_mh_data(
    data = colo_site_set_rates
    )
}

}
