% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_change_cumu.R
\name{calc_change_cumu}
\alias{calc_change_cumu}
\title{Calculate station- or site-level cumulative change for SET or MH data}
\usage{
calc_change_cumu(data, level = "station")
}
\arguments{
\item{data}{A data frame. A data frame of raw SET or MH data. See details
below for requirements.}

\item{level}{A string (optional). Level at which to calculate rates of
surface elevation change or vertical accretion. One of:
* `"station"`: (default) station-level rates of surface elevation change.
* `"site"`: site-level rates of surface elevation change.}
}
\value{
For SET data, returns a data frame of station- or site-level
  cumulative surface elevation change. For MH data, returns a data frame of
  station- or site-level vertical accretion.
}
\description{
This function takes a data frame of raw SET or MH data and calculates
cumulative surface elevation change (for SET data) or vertical accretion (for
MH data) on each measurement date at either the station- or site-level.
}
\section{Data Requirements}{


  This function takes a data frame of either raw SET or MH data (`data`).

  SET data must have 1 row per pin reading and the following columns, named
  exactly: event_date_UTC, network_code, park_code, site_name, station_code,
  SET_direction, pin_position, SET_offset_mm, pin_length_mm, and
  pin_height_mm. Note that SET_offset_mm and pin_length_mm can be empty (aka
  blank) but the columns must be included in the data frame. See
  `example_sets`.

  MH data must have 1 row per core measurement and the following columns,
  named exactly: event_date_UTC, network_code, park_code, site_name,
  marker_horizon_name, core_measurement_number, core_measurement_depth_mm,
  and established_date. See `example_mh`.
}

\section{Details}{


  For SET data, pin-level cumulative change is first calculated as the
  difference between each pin reading and the reading from the earliest date
  that was not NA. Note that for data from NCBN, NCRN, and NETN, pin heights
  are first converted to standardized pin heights using the formula 1000 +
  (SET_offset_mm - (pin_length_mm - pin_height_mm)) to account for the 6"
  extensions that were added to some benchmarks within these networks' parks
  in 2023/24. Then, cumulative pin changes are averaged to the arm-level on
  each date, and then the cumulative arm-level changes are averaged up to the
  station-level on each date. If level = "site" is used, station-level
  cumulative change will be averaged up to the site-level.

  For MH data, duplicate core readings ('core_measurement_number') are first
  averaged to the marker horizon-level. Then, average marker horizon depths
  are averaged up to the station-level on each date, excluding NAs. If level
  = "site" is used, station-level cumulative change will be averaged up to
  the site-level.
}

\examples{
# Defaults to station-level cumulative change
calc_change_cumu(example_sets)

# Site-level cumulative change
calc_change_cumu(example_mh, level = "site")

# Can also be used with pipes to include/exclude specific stations, dates, etc.
## Exclude data from after 2016 from calculations of cumulative change
example_sets \%>\%
    filter(event_date_UTC < as.Date("2016-01-01")) \%>\%
    calc_change_cumu(., level = "station")

## Define custom groups for calculating cumulative change
example_sets \%>\%
    group_by(set_type) \%>\%
    calc_change_cumu(., level = "site")

}
