% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_linear_rates.R
\name{calc_linear_rates}
\alias{calc_linear_rates}
\title{Calculate station- or site-level linear rates of change from SET or MH data}
\usage{
calc_linear_rates(data, level = "station")
}
\arguments{
\item{data}{A data frame. A data frame of raw SET or MH data. See details
below for requirements.}

\item{level}{A string (optional). Level at which to calculate rates of
surface elevation change or vertical accretion. One of:
* `"station"`: (default) station-level rates of surface elevation change.
* `"site"`: site-level rates of surface elevation change.}
}
\value{
For SET data, returns a data frame of station- or site-level rates of
  surface elevation change. For MH data, returns a data frame of station- or
  site-level rates of surface accretion. Note that units for rates of change
  are mm/yr.
}
\description{
This function takes a data frame of raw SET or MH data and calculates a
linear rate of surface elevation change (for SET data) or vertical accretion
(for MH data).
}
\note{
Stations with fewer than 3 measurement dates are excluded from the
  calculation of rates. Cumulative change is calculated via the function
  [calc_change_cumu] - see function documentation for details.
}
