% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_set_data.R
\name{load_set_data}
\alias{load_set_data}
\title{Calculate Cumulative Change at a SET}
\usage{
load_set_data(park_code = NULL, network_code = NULL, file_path = NULL)
}
\arguments{
\item{park_code}{optional; A 4 character park code that can be used to filter results from the NPS SET database to a specific park.}

\item{network_code}{optional; A 4 character network code that can be used to filter results from the NPS SET database to a specific I&M network.}

\item{file_path}{optional; If, instead of connecting to the NPS SET database, you want to use a saved .csv or .xlsx file, the full file path to the saved file. Must be table with one row per pin reading, and the following columns, named exactly: event_date_UTC, network_code, park_code, site_name, station_code, SET_direction, pin_position, pin_height_mm.}
}
\value{
A data frame of raw SET data.
}
\description{
Calculate Cumulative Change at a SET
}
\examples{

load_set_data(data_type = "database")

load_set_data(data_type = "saved_file", file_path = "C:/Documents/Data/my_SET_data.csv")

}
