% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_set_data.R
\name{load_set_data}
\alias{load_set_data}
\title{Calculate Cumulative Change at a SET}
\usage{
load_set_data(data_type = "database", file_path = NULL, park_code = NULL)
}
\arguments{
\item{data_type}{Either "database" if you want to connect to the NPS SET database (authorized users) or "saved_file" if you have a saved .csv or xlse file.}

\item{file_path}{If using a saved .csv or .xlsx file, the full file path to the saved file. Must be table with one row per pin reading, and the following columns, named exactly: event_date_UTC, network_code, park_code, site_name, station_code, SET_direction, pin_position, pin_height_mm.}

\item{park_code}{A 4 character park code that can be used to filter results from the NPS database to a specific park.}
}
\value{
A data frame of raw SET data.
}
\description{
Calculate Cumulative Change at a SET
}
\examples{

load_set_data(data_type = "database")

load_set_data(data_type = "saved_file", file_path = "C:/Documents/Data/my_SET_data.csv")

}
