% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumu.R
\name{plot_cumu}
\alias{plot_cumu}
\title{Plot station- or site-level cumulative surface elevation change or vertical
accretion}
\usage{
plot_cumu(
  SET_data = NULL,
  MH_data = NULL,
  level = "station",
  rate_type = NULL,
  columns = 4,
  pointsize = 2,
  scales = "fixed"
)
}
\arguments{
\item{SET_data}{A data frame. A data frame of raw SET data. See details below
for requirements. If both SET and MH data are supplied, plots both.}

\item{MH_data}{A data frame. A data frame of raw MH data. See details below
for requirements. If both SET and MH data are supplied, plots both.}

\item{level}{A string (optional). Level at which to calculate rates of
surface elevation change and/or vertical accretion. One of:
* `"station"` (default) station-level rates of surface elevation change.
* `"site"` site-level rates of surface elevation change.}

\item{rate_type}{A string (optional); Calculate linear rate of change and
include in the plot? Use rate_type = `"linear"` to display station- (level
= `"station"`) or site-level (level = `"site"`) rates of surface elevation
change or vertical accretion in mm/yr.}

\item{columns}{An integer. Number of columns you want in the faceted output;
defaults to 4. Utilizes the `ncol` argument of [ggplot2::facet_wrap].}

\item{pointsize}{A number. Size of points on the plot; defaults to 3.5.
Utilizes the `size` argument of [ggplot2::geom_point].}

\item{scales}{Do you want axis scales to be the same in all facets ("fixed")
or to vary between facets? Defaults to using the same scale between all
panels. Utilizes the `scales` parameter of [ggplot2::facet_wrap]. One of:
* `"fixed"`: default; use the same axis scales in all panels.
* `"free_x"`: allow x-axis scales to vary between panels.
* `"free_y"`: allow y-axis scales to vary between panels.
* `"free"`: allow both x- and y-axis scales to vary between panels.}
}
\value{
A ggplot object: x-axis is date; depending on the type of data
  supplied, y-axis is either (a) cumulative surface elevation change and/or
  (b) vertical accretion. level = "station" returns a plot with one panel per
  station, whereas level = "site" returns a plot with one panel per site.
}
\description{
Creates a faceted plot of cumulative surface elevation change and/or vertical
accretion at each station or site. Optionally calculates and displays linear
rates of change.
}
\note{
Cumulative change is calculated via the function [calc_change_cumu] and
  linear rates of change are calculated via the function [calc_linear_rates].
  See function documentation for details.
}
\examples{
# Station-level cumulative change
plot_cumu(SET_data = example_sets)

# Combined plot of site-level SET and MH cumulative change
plot_cumu(SET_data = example_sets, MH_data = example_mh, level = "site")

# Change the number of columns, point size, or scaling of x- and/or y-axis
plot_cumu(SET_data = example_sets, MH_data = example_mh, columns = 2, pointsize = 1, scales = "free")

# Apply a date filter to both SET and MH, and then plot
df_list <- list("SET" = example_sets, "MH" = example_mh) \%>\%
map(., ~.x \%>\%
       filter(event_date_UTC < as.Date("2016-01-01")) \%>\%
       nest()) \%>\%
       list_cbind()

# Modify the 'station_code' or 'site_name' columns to facet plots by custom groups
plot_cumu(SET_data = unnest(df_list$SET, cols = "data"), MH_data = unnest(df_list$MH, cols = "data"))

example_sets_set_type <- example_sets \%>\%
    mutate(site_name = if_else(station_code == "M11-2", paste0(site_name, " shallow"), paste0(site_name," deep")))

example_mh_set_type <- example_mh \%>\%
    mutate(site_name = if_else(station_code == "M11-2", paste0(site_name, " shallow"), paste0(site_name," deep")))

plot_cumu(SET_data = example_sets_set_type, MH_data = example_mh_set_type, level = "site", rate_type = "linear")

}
