% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rate_comps.R
\name{plot_rate_comps}
\alias{plot_rate_comps}
\title{Graphical comparison of station- or site-level surface elevation change rates}
\usage{
plot_rate_comps(data = NULL, rates = NULL, level = "station")
}
\arguments{
\item{data}{A data frame (optional). A data frame of raw SET data. See
details below for requirements.}

\item{rates}{A data frame (optional). A data frame of station- or site-level
rates of surface elevation change. See details below for requirements.}

\item{level}{A string (optional). Level at which to calculate rates of
surface elevation change. One of:
* `"station"`: (default) station-level rates of surface elevation change.
* `"site"`: site-level rates of surface elevation change.}

\item{data_type}{A string (required). Is the data supplied to `data` or
`rates` SET or MH data? One of:
* `"SET"`: SET data.
* `"MH"`: MH data.}
}
\value{
a ggplot object
}
\description{
This function creates a 'TIE fighter' style plot that can be
  used to visually compare station- or site-level linear rates of surface
  elevation change.

  Can use a data frame of either raw SET data (`data` argument) or a data
  frame of pre-calculated station- or site-level rates of surface elevation
  change (`rates` argument).
}
\section{Data Requirements}{


  This function takes a data frame of either raw SET data (`data`) or a
  user-created data frame of station- or site-level rates of surface
  elevation change (`rates`).

  If supplying `data` and `data_type = "SET"`, raw SET data must have 1 row
  per pin reading and the following columns, named exactly: event_date_UTC,
  network_code, park_code, site_name, station_code, SET_direction,
  pin_position, SET_offset_mm, pin_length_mm, and pin_height_mm. Note that
  SET_offset_mm and pin_length_mm can be NA (aka blanks) but the columns must
  be included in the data frame.

  If supplying `data` and `data_type = "MH"`, raw MH data must have 1 row per
  core measurement and the following columns, named exactly: event_date_UTC,
  network_code, park_code, site_name, station_code, marker_horizon_name,
  core_measurement_number, core_measurement_depth_mm, and established_date

  If supplying a user created data frame to  `rates` and `level = station`,
  must be a data frame of station-level rates of surface elevation change or
  vertical accretion with 1 row per station and the following columns, named
  exactly: station_code, rate, and rate_se. The "rate" and "rate_se" columns
  should represent the station-level rates and their standard errors
  (respectively).

  If supplying a user created data frame to `rates` and `level = site`, must
  be a data frame of site-level rates of surface elevation change or vertical
  accretion with 1 row per site and the following columns, named exactly:
  site_name, rate, and rate_se. The "rate" and "rate_se" columns should
  represent the site-level rates and their standard errors (respectively).
}

\section{Note}{


  Cumulative change is calculated via the function [calc_change_cumu] and
  linear rates of change are calculated via the function [calc_linear_rates].
  See function documentation for details.
}

\examples{

# Default plot gives station-level rates
plot_set_rate_comps(data = example_sets)

# Example with a user-supplied data frame of station-level rates
example_rates <- data.frame("station" = c("station_1", "station_2", "station_3"),
                            "rate" = c(3.2, 4.0, 5.4),
                            "se_rate" = c(1, 0.5, 0.25))

plot_set_rate_comps(rates = example_rates)

# Site-level rates
plot_set_rate_comps(data = example_sets, level = "site")

}
