% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rate_comps.R
\name{plot_rate_comps}
\alias{plot_rate_comps}
\title{Graphical comparison of station- or site-level surface elevation change rates}
\usage{
plot_rate_comps(data = NULL, rates = NULL, level = "station")
}
\arguments{
\item{data}{A data frame (optional). A data frame of raw SET data. See
details below for requirements.}

\item{rates}{A data frame (optional). A data frame of station- or site-level
rates of surface elevation change. See details below for requirements.}

\item{level}{A string (optional). Level at which to calculate rates of
surface elevation change. One of:
* `"station"`: (default) station-level rates of surface elevation change.
* `"site"`: site-level rates of surface elevation change.}

\item{data_type}{A string (required). Is the data supplied to `data` or
`rates` SET or MH data? One of:
* `"SET"`: SET data.
* `"MH"`: MH data.}
}
\value{
a ggplot object
}
\description{
This function creates a 'TIE fighter' style plot that can be
  used to visually compare station- or site-level linear rates of surface
  elevation change.

  Can use a data frame of either raw SET data (`data` argument) or a data
  frame of pre-calculated station- or site-level rates of surface elevation
  change (`rates` argument).
}
\examples{

# Default plot gives station-level rates
plot_set_rate_comps(data = example_sets)

# Example with a user-supplied data frame of station-level rates
example_rates <- data.frame("station" = c("station_1", "station_2", "station_3"),
                            "rate" = c(3.2, 4.0, 5.4),
                            "se_rate" = c(1, 0.5, 0.25))

plot_set_rate_comps(rates = example_rates)

# Site-level rates
plot_set_rate_comps(data = example_sets, level = "site")

}
